+++ dhcpkit/ipv6/extensions/todo/prefix_delegation_exclude.py
@@ -1,3 +1,0 @@
-# http://www.iana.org/go/rfc6603
-
-OPTION_PD_EXCLUDE = 67